build: update libkcoidc support to use pkg-config
Newer versions of libkcoidc gained support for pkg-config. This
change updates the corresponding detection to require libkcoid
version 0.9.0 or later.

Furthermore also the `--enable-kcoidc` configure parameter to require
libkcoidc is added. If that parameter is not given, the build is
configured with libkcoic if found and otherwise without. This makes
it possible for packaging systems to ensure that kcoidc is properly
supported.

(cherry picked from commit ea1995d3d5c70b9448b5752a138ad9e4824345a7)