test: require libkustomer when running test suite
To ensure that we test supported builds, this change enables
the kustomer configure flag to ensure that compile happens with
libkustomer enabled. For the time being, kustomerd is not run
in the test suite and thus the storage server is allowed to start
without waiting on kustomer initialization by using 0 as
kustomer initialization timeout.