libserver: respond to getLicenseAuth RPC in original fashion
Re-implement responding to the getLicenseAuth RPC with full frames.
Since the logon RPC happens all the time and is not indicative of the
program that sits on top of libmapi, the getLicenseAuth RPC is used
to carry out program-specific license checks.

References: KC-1821
(cherry picked from commit 9ff5c3da211156119a7d9844490210322d879639)