test: add support to install extra packages
To test with libkustomer, the library package needs to be
installed. While the test suite already had the option to add
extra packages while building, it was lacking the same option
for the containers which actually run. This change adds the
option via build args to add extra packages and uses it to
install libkustomer0 into the kopano_server container when run
via Jenkins.