unixplugin: drop use of getpwent_r
UnixUserPlugin.cpp:355:7: error: "getpwent_r" was not declared in this
scope; did you mean "getpwent"?

There is not a lot of sense in getpwent_r. Though it avoids the use
of a singleton buffer, it still operates on a singleton iterator
that is used to walk the NSS database.

References: KF-3361