Dockerfile.build: fix CONFIGURE_ARGS support
Docker build args are not available directly in the CMD command since
that is executed at runtime. Thus any CONFIGURE_ARGS value was not used
when running the Docker based build with its default command. This
change exports the build arg value to the container environment, making
it available to the default command at runtime.