common: drop "When reporting" guideline
Users don't read documentation, and this extends to logs. Many
reports in the past failed to come with the requested extra
information, up to a breaking point when ECLogger was made to emit
the OS information automatically. Thenceforth, the user need not take
any extra action and so that line of request can be dropped.