pyko: disable syslog log method support
The syslog log method does not work correctly and thus this change
disables it. Until it is fixed, Python services do not support
the syslog log_method. Additionally an environment variable is
added, which allow to lift this restriction for testing when it
it set.

Reference: KC-1858
(cherry picked from commit 4a7f833e170167ebfa4f4c55835f8760ce7617f3)