pyko: honor log_timestamp log setting
Sometimes it is not useful to log time stamps by the service. For
this the configuration option log_timestamp is available but it
was not used by pyko services. This change honors the value and
also updates the log format to match how the non-Python services
of this project.

Fixes: KC-1862
(cherry picked from commit b26459f264a24ecba55ab973ad8729d371264619)