archiver: trim EC prefix
The symbols were put into "namespace KC" around the time of commit
bp/8.3.0-209-g2d60ab389, which makes the lettered prefix serve no
more purpose. This reduces the size of libraries somewhat, as all
the symbol names get shorter (and there is generally a lot of those
with GNU C++).