ldapplugin: reduce time to obtain DN cache
A dn_cache map with some 37700 entries takes around 800000 µsec to
copy; during this time, the cache mutex is also held. The copy made
by getObjectDNCache is not warranted for how the data is used
afterwards. Return a reference instead. New observed hold times are
now 500-1200 µsec on that same system.

Fixes: SVN r13598
References: KC-1876