treewide: declare ECMemStream::Seek return value for ignore
CID-215525 (#1 of 1): Unchecked return value (CHECKED_RETURN)
CID-215523 (#1 of 1): Unchecked return value (CHECKED_RETURN)
1. check_return: Calling Seek without checking return value
(as is done elsewhere 8 out of 10 times).

When seek to zero fails (IStream API), there is nothing meaningful to
do about it. In practice, the only time that I can imagine when this
would happen is if the stream is a pipe. Kopano only has ECMemStream
(not a pipe), so never fails.