common: rewrite fd closing in unix_popen_rw for cov-scan
open_fn: Returning handle opened by open.
5. var_assign: Assigning: nullfd = handle returned from open("/dev/null", 2).
CID-210503 (#1 of 1): Resource leak (RESOURCE_LEAK)
8. leaked_handle: Handle variable nullfd going out of scope
leaks the handle.

The handle does not leak, because of the cl1 function object. It
would appear Coverity does not see through that, so move cl1's
behavior to a non-lambda class with classic destructor in the hope
of resolving 210503.