client: attach buffers used for soap requests to the request
Allocate buffers using the soap_new_*T* calls. This not only invokes
the constructor for Ts (which so far has been missing), but also
attaches the buffers to the soap request, so memory management is
automatic (m_lpCmd->soap gets cleared/reset after each request), and
we will not need to invoke FreeABProps anymore.