provider: rectify erroneous HRESULT-ECRESULT mixing
cmd.cpp had this pattern in one place:

	er = MAPI...
	if (er != erSuccess)
		goto exitm;
	...
	exitm:
	return hr;

This is wrong for two reasons: Most functions by the name of /^MAPI/ return
HRESULT, while er is declared as an ECRESULT; furthermore, this result was not
returned up the call chain. This could make kopano-server respond to RPCs with
a success code even though the operation failed (hid errors).

Scanned the source for

	/er = MAPI/
	/hr .= er/

and corrected the variable names or return conventions.