build: cure printf size warning for libfreebusy
libfreebusy/recurrence.cpp: In member function 'void KC::recurrence::HrGetHumanReadableString(std::string*)':
libfreebusy/recurrence.cpp:131:45: warning: '%02u' directive output may be truncated writing 2 bytes into a region of size between 0 and 5 [-Wformat-truncation=]
  131 |  snprintf(startocc, sizeof(startocc), "%02u:%02u", getStartTimeOffset() / 3600, getStartTimeOffset() / 60 % 60);

getStartTimeOffset is not supposed to return a value larger than 86400. The compiler
does not know this and assumes the worst, in which case the first %02u would fill
most of the 8-byte buffer and leave too little for the second %02u (hence the warning).