pyko: Set PR_FREEBUSY_ENTRYIDS correctly for a MSR store
Python-kopano has a create store method for MSR due to limitations in
the C++ CreateStore method, this however has a creates the
PR_FREEBUSY_ENTRYIDS incorrectly. When the calendar permissions are set
on the new store, the "FreeBusy Data" folder also needs to be
visible/write-able for the delegate, but in the initial version this
folder is referenced as '' which opens the root folder. This creates a
security issue as it sets extranous read/write permissions on the root
folder instead of the 'Freebusy Data' folder.