ParseICal reports failed components and properties
[KC-1852] After parsing the ical string two member variables are set to
indicate if any errors occured during parsing.
The errors essentially can boil down to an invalid property or a
component that fails to parse. This can happen, for example, if the
component does not have the minimum required properties.
An invalid property that results in an invalid component will result in
both variables being incremented.