Do not use local time when storing all day events
According to the spec all day events should be treated as DATE only in
UTC. Our previous implementation was converting the date to local time.
Afterwards some guess work was being done that did not always work.
For example for timezones set in GMT+14 the incorrect day was
being calculated.

References: KC-1664