Compute normalized subject based on prefix prop
The normalized subject should be computed based on what is present in
the PR_SUBJECT_PREFIX property. This also fixes an issue that occured
when character encodings that are larger than 1 byte were stored as a
non wide char. The previous code was counting the bytes, always
assuming that 1 byte = 1 character. That is not true with, for example,
japanese characters.
Note: An explanation as to why we don't 100% follow the MAPI
documentation on the normalized subject was also added as a comment.