libicalmapi: fix incomplete conversion from wchar_t to std::string
The BDAY and X-ANNIVERSARY field were never parsed correctly as the
conversion of whcar_t to std::string returned a nullptr. A new functions
is introduced to convert a wchar_t to char* for use of
date_string_to_filetime. The second issue is an exported VCF file
returns a date string which we do not support yet.

References: KC-1608