Add config headers to inhibit auto reply and forwd
This creates two new configurations for users to add a list of headers
that inhibit auto replying and auto forwarding e-mails.
Caveat: Previous algorithm was using a hash map which provided better
complexity. To simplify the algorithm this is now done with vectors
however I do not expect it to have any noticable impact due to them
have a relatively small size. We can always improve it if necessary.
References: KC-1525