Add a function to create a new iconv_context
Previously we had no way to create a new context and check if the
charsets are compatible unless we try to convert. This allows a user
of the class to create a context in advance and catch exceptions
(typically due to invalid charsets). A function to do the same on the
global context was also added.