Close and re-open iconv object before reusing it
According to iconv's documentation passing a nullptr as a buffer should
reset it to its intial state. However it seems sometimes we get an
unexpected conversion result. Running the unittests reveals this
problem.
Example: when converting from UTF-32LE to //TRANSLIT encoding,
from a wide character to a character.
If the original string is ザラファ we expect to get ????, since we shouldn't
be able to represent these characters. However, if we don't close and
re-open the iconv object we get the correct value back: ザラファ
As of now I haven't been able to figure out why this happens, so to move
on, this solution is in place. A comment has been left so this can be
revisited if necessary.