Add PHP 8 compiliation support
Update vendored run-tests to PHP 8 version, add argument information as
this is now a requirement for PHP 8 compatibility. Drop all TSRMLS_*
constants which are already deprecated and only relevant for PHP 5 which
is no longer supported.

Fix invalid conversion from ‘int’ to ‘zend_result. PHP_MINIT_FUNCTION
expects a zend_result as return value, not int. This is now an compiler
error in PHP 8.