pyko: handle ICS race condition when folder is deleted
When an item is first created in a folder which is immediately deleted a
race condition can occur in checking if the folder exists. Assigning a
local variable circumvents this issue, an event for a deleted folder
might still go through but the "ICS consumer" should handle this.

References: KC-1997